Add solution for problem `B - Running for Gold` Link: https://codeforces.com/contest/1552/problem/B Tags: combinatorics, graphs, greedy, sortings, *1500 Ref: https://codeforces.com/contest/1552/submission/123803184

+++ codeforces/1552/B.cpp
@@ -65,7 +65,7 @@ bool isSup(mat &a, int i, int j) {
 
 void solve()
 {
-      ll t1 = 1;
+      int t1 = 1;
       cin >> t1;
       for (ll tt = 1; tt <= t1; tt++) {
             int n; cin >> n;
